+++ src/main/java/com/chaosbuffalo/mkultra/client/gui/lib/MKWidget.java
@@ -95,7 +95,7 @@ public class MKWidget extends Gui {
         return this;
     }
 
-    public Vec2d getParentCoords(Vec2d pos) {
+    public Vec2i getParentCoords(Vec2i pos) {
         if (parent == null) {
             return pos;
         } else {
