+++ src/main/java/com/chaosbuffalo/mkultra/client/gui/lib/Vec2i.java
@@ -1,27 +1,27 @@
 package com.chaosbuffalo.mkultra.client.gui.lib;
 
-public class Vec2d {
+public class Vec2i {
     public int x;
     public int y;
 
-    public Vec2d(int x, int y) {
+    public Vec2i(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public Vec2d subtract(Vec2d other) {
+    public Vec2i subtract(Vec2i other) {
         return this.subtract(other.x, other.y);
     }
 
-    public Vec2d subtract(int xVal, int yVal) {
+    public Vec2i subtract(int xVal, int yVal) {
         return this.add(-xVal, -yVal);
     }
 
-    public Vec2d add(Vec2d other) {
+    public Vec2i add(Vec2i other) {
         return this.add(other.x, other.y);
     }
 
-    public Vec2d add(int xVal, int yVal) {
-        return new Vec2d(x + xVal, y + yVal);
+    public Vec2i add(int xVal, int yVal) {
+        return new Vec2i(x + xVal, y + yVal);
     }
 }
